Split mixup calculation logic to functions

+++ mixup/ood_score_calculators/mixdiff_energy.py
@@ -14,7 +14,9 @@ class MixDiffEnergy(
         known_logits,
         unknown_logits,
     ):
-        logit_scale = self.clip_model.logit_scale.exp()
+        logit_scale = self.clip_model.logit_scale.exp().float()
+        known_logits = known_logits.float()
+        unknown_logits = unknown_logits.float()
         known_neg_energy = torch.logsumexp(known_logits, dim=-1) / logit_scale
         unknown_neg_energy = torch.logsumexp(unknown_logits, dim=-1) / logit_scale
         mixdiff = -(unknown_neg_energy - known_neg_energy)
@@ -27,7 +29,8 @@ class MixDiffEnergy(
         **kwargs,
     ):
         if self.add_base_score:
-            logit_scale = self.clip_model.logit_scale.exp()
+            logit_scale = self.clip_model.logit_scale.exp().float()
+            logits = logits.float()
             neg_energy = torch.logsumexp(logits, dim=-1) / logit_scale
         else:
             neg_energy = torch.zeros_like(logits[..., -1])
